Изменение в обратной связи и оптимизации
