+++ js/util.js
@@ -18,48 +18,6 @@ const getRandomArrayElement = (elements) => elements[getRandomInt(0, elements.le
 
 export {getRandomInt, getRandomFloat, getRandomArrayElement};
 
-const isEscapeKey = (evt) => evt.key === 'Escape';
-
-export {isEscapeKey};
-
-const ALERT_SHOW_TIME = 5000;
-
-const showAlert = (message) => {
-  const alertContainer = document.createElement('div', { is : 'show-alert' });
-  alertContainer.textContent = message;
-  document.body.append(alertContainer);
-
-  setTimeout(() => {
-    alertContainer.remove();
-  }, ALERT_SHOW_TIME);
-};
-
-export {showAlert};
-
-const showMessage = (messageType) => {
-  const typeMessageTamplate = document.querySelector(`#${messageType}`).content.querySelector(`.${messageType}`);
-  const massage = typeMessageTamplate.cloneNode(true);
-  document.querySelector('.notice').append(massage);
-  document.addEventListener('keydown', (evt) => {
-    if (isEscapeKey(evt)) {
-      evt.preventDefault();
-      massage.classList.add('hidden');
-    }
-  });
-  document.addEventListener('click', () => {
-    massage.classList.add('hidden');
-  });
-
-  const closeButton = typeMessageTamplate.querySelector(`.${messageType}__button`);
-  if (closeButton) {
-    closeButton.addEventListener('click', () => {
-      massage.classList.add('hidden');
-    });
-  }
-};
-
-export {showMessage};
-
 function debounce (callback, timeoutDelay) {
   let timeoutId;
   return (...rest) => {
